fix: resolve 20 cpplint errors blocking CI preflight
- Add missing newlines at EOF (4 files)
- Fix header guards (2 files)
- Fix namespace closing comment (context.hpp)
- Replace unsigned long long / long long with uint64_t / int64_t (10 lines)
- Replace using-directive with using-declarations (partition_config.cc)
- ASSERT_TRUE → ASSERT_EQ/ASSERT_GE (2 test files)

cpplint: 0 errors (was 20). Build: clean (GCC 15.2.0, zero warnings).
89/89 CDC+vector+partition tests pass.

+++ src/serializer/logical_journal.cc
@@ -314,7 +314,7 @@ void logical_journal_t::create(UNUSED signal_t *interruptor) {
         if (index_exists) {
             if (st.st_size > 0 && st.st_size < static_cast<off_t>(INDEX_HEADER_SIZE)) {
                 crash("logical_journal_t::create: index file %s is truncated (size=%lld)",
-                      idx_path_.c_str(), static_cast<long long>(st.st_size));
+                      idx_path_.c_str(), static_cast<int64_t>(st.st_size));
             }
         } else {
             int err = 0;
